requirements-dev.txt
-r requirements.txt
black
pylint
flake8
django-debug-toolbar
mypy